statistics: move MergePartTopN2GlobalTopN into handle/globalstats
