+++ pkg/statistics/cmsketch.go
@@ -23,8 +23,6 @@ import (
 	"slices"
 	"sort"
 	"strings"
-	"sync/atomic"
-	"time"
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
@@ -793,84 +791,6 @@ func (c *TopN) updateTopNWithDelta(d []byte, delta uint64, increase bool) bool {
 // NewTopN creates the new TopN struct by the given size.
 func NewTopN(n int) *TopN {
 	return &TopN{TopN: make([]TopNMeta, 0, n)}
-}
-
-// MergePartTopN2GlobalTopN is used to merge the partition-level topN to global-level topN.
-// The input parameters:
-//  1. `topNs` are the partition-level topNs to be merged.
-//  2. `n` is the size of the global-level topN. Notice: This value can be 0 and has no default value, we must explicitly specify this value.
-//  3. `hists` are the partition-level histograms. Some values not in topN may be placed in the histogram. We need it here to make the value in the global-level TopN more accurate.
-//
-// The output parameters:
-//  1. `*TopN` is the final global-level topN.
-//  2. `[]TopNMeta` is the left topN value from the partition-level TopNs, but is not placed to global-level TopN. We should put them back to histogram latter.
-//  3. `[]*Histogram` are the partition-level histograms which just delete some values when we merge the global-level topN.
-func MergePartTopN2GlobalTopN(loc *time.Location, version int, topNs []*TopN, n uint32, hists []*Histogram,
-	isIndex bool, killed *uint32) (*TopN, []TopNMeta, []*Histogram, error) {
-	if CheckEmptyTopNs(topNs) {
-		return nil, nil, hists, nil
-	}
-	partNum := len(topNs)
-	// Different TopN structures may hold the same value, we have to merge them.
-	counter := make(map[hack.MutableString]float64)
-	// datumMap is used to store the mapping from the string type to datum type.
-	// The datum is used to find the value in the histogram.
-	datumMap := NewDatumMapCache()
-	for i, topN := range topNs {
-		if atomic.LoadUint32(killed) == 1 {
-			return nil, nil, nil, errors.Trace(ErrQueryInterrupted)
-		}
-		if topN.TotalCount() == 0 {
-			continue
-		}
-		for _, val := range topN.TopN {
-			encodedVal := hack.String(val.Encoded)
-			_, exists := counter[encodedVal]
-			counter[encodedVal] += float64(val.Count)
-			if exists {
-				// We have already calculated the encodedVal from the histogram, so just continue to next topN value.
-				continue
-			}
-			// We need to check whether the value corresponding to encodedVal is contained in other partition-level stats.
-			// 1. Check the topN first.
-			// 2. If the topN doesn't contain the value corresponding to encodedVal. We should check the histogram.
-			for j := 0; j < partNum; j++ {
-				if atomic.LoadUint32(killed) == 1 {
-					return nil, nil, nil, errors.Trace(ErrQueryInterrupted)
-				}
-				if (j == i && version >= 2) || topNs[j].FindTopN(val.Encoded) != -1 {
-					continue
-				}
-				// Get the encodedVal from the hists[j]
-				datum, exists := datumMap.Get(encodedVal)
-				if !exists {
-					d, err := datumMap.Put(val, encodedVal, hists[0].Tp.GetType(), isIndex, loc)
-					if err != nil {
-						return nil, nil, nil, err
-					}
-					datum = d
-				}
-				// Get the row count which the value is equal to the encodedVal from histogram.
-				count, _ := hists[j].EqualRowCount(nil, datum, isIndex)
-				if count != 0 {
-					counter[encodedVal] += count
-					// Remove the value corresponding to encodedVal from the histogram.
-					hists[j].BinarySearchRemoveVal(TopNMeta{Encoded: datum.GetBytes(), Count: uint64(count)})
-				}
-			}
-		}
-	}
-	numTop := len(counter)
-	if numTop == 0 {
-		return nil, nil, hists, nil
-	}
-	sorted := make([]TopNMeta, 0, numTop)
-	for value, cnt := range counter {
-		data := hack.Slice(string(value))
-		sorted = append(sorted, TopNMeta{Encoded: data, Count: uint64(cnt)})
-	}
-	globalTopN, leftTopN := GetMergedTopNFromSortedSlice(sorted, n)
-	return globalTopN, leftTopN, hists, nil
 }
 
 // MergeTopN is used to merge more TopN structures to generate a new TopN struct by the given size.
