+++ pkg/statistics/handle/globalstats/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "global_stats.go",
         "global_stats_async.go",
+        "merge_topn.go",
         "merge_worker.go",
         "topn.go",
     ],
@@ -39,10 +40,10 @@ go_test(
         "main_test.go",
         "topn_bench_test.go",
     ],
+    embed = [":globalstats"],
     flaky = True,
     shard_count = 19,
     deps = [
-        ":globalstats",
         "//pkg/config",
         "//pkg/parser/model",
         "//pkg/parser/mysql",
